==> Part 4 - More cleaning and merging candidate df with tweet df.ipynb <==
import pandas as pd

data = pd.read_csv('pres.csv')
tweets = pd.read_csv('pres_tweets.csv', index_col = 0)

data.dtypes

tweets.candidate.value_counts(dropna=False)

#change date column to datetime
tweets['date'] = pd.to_datetime(tweets['date'], errors='coerce')

#drop null
tweets.dropna(inplace=True)
tweets.reset_index(drop=True, inplace=True)

tweets.isnull().sum()

tweets['year'] = tweets['date'].dt.year

tweets.head(5)

#merge df on candidate and year
tweets_merged = pd.merge(tweets, data,  how='left', left_on=['candidate','year'], right_on = ['candidate','year'])


tweets_merged.head(5)

tweets_merged.to_csv('pres_tweets_merged.csv', index=False)

# merge gubernatorial df

g_cand = pd.read_csv('gub.csv', index_col=0)
g_tweets = pd.read_csv('gub_tweets_processed.csv', index_col = 0)

g_cand['party'].value_counts()

def clean_party(x):
    if x in ['Democratic', 'D', 'DFL']:
        return 'Democratic'
    elif x in ['Republican', 'R']:
        return 'Republican'
    else:
        return 'Other'

g_cand['party'] = list(map(clean_party, g_cand['party']))

g_tweets['date'] = pd.to_datetime(g_tweets['date'], errors='coerce')

g_tweets.isnull().sum()

g_tweets.dropna(inplace=True)
g_tweets.reset_index(drop=True, inplace=True)

g_tweets.head(5)

g_tweets['year'] = g_tweets['date'].dt.year

#merge df on candidate and year
g_tweets_merged = pd.merge(g_tweets, g_cand,  how='left', left_on=['candidate','year'], right_on = ['candidate','year'])

g_tweets_merged['percent_popular_vote'] = g_tweets_merged['percent_popular_vote']/100

g_tweets_merged.head(5)

g_tweets_merged.to_csv('gub_tweets_merged.csv', index=False)
