==> Part 7 - building outcome prediction model.ipynb <==
import pandas as pd

data = pd.read_csv('gub_candidates.csv')
data

#clean party column
def clean_party(x):
    party_dict = {'D':'Democratic', 'DFL':'Democratic', 'R':'Republican', 'L':'Libertarian', 'I':'Independent', 'ndependent':'Independent', 'G':'Green','C':'Constitution'}
    if x in ['Republican', 'Democratic', 'Libertarian', 'Independent', 'Green', 'Constitution']:
        return x
    elif x in party_dict.keys():
        return party_dict[x]
    else:
        return 'Other'

data['party'] = list(map(clean_party, data['party']))

data['party'].value_counts()

data['winner'] = data['winner'].astype(object)

## X-y split

y1 = data['winner']
y2 = data['percent_popular_vote']
X = data.drop(['winner', 'percent_popular_vote'], axis=1, inplace=True)

## normalize features

#split numericals and categoricals
X_num = data[['percent_positive', 'percent_total_tweets']]
X_cat = data[['year','state', 'party']]

#OneHotEncode the categoricals
from sklearn.preprocessing import OneHotEncoder
encoder = OneHotEncoder(drop='first')
encoded = encoder.fit_transform(X_cat).toarray()
Xcat_encoded = pd.DataFrame(encoded)
Xcat_encoded.columns = encoder.get_feature_names()

X_normalized = pd.concat([X_num, Xcat_encoded], axis=1)

X_normalized

## making model for predicting percent popular vote

The train-test split needs to be done by election so that it is possible to compare %popular vote to predict the winner. Therefore, I am using the 2018 elections as the test set as it is 111 candidates, or 22% of the total data. 

data[data['year']==2018]

## train-test split
X_test = X_normalized[45:156]
X_train = pd.concat([X_normalized[:45], X_normalized[156:]], axis=0)
y_test= y2[45:156]
y_train = pd.concat([y2[:45], y2[156:]], axis=0)

len(y_train)

## make model pipelines

import numpy as np
import math
from sklearn.linear_model import LinearRegression
from sklearn.tree import DecisionTreeRegressor
from sklearn.neighbors import KNeighborsRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import cross_val_score
from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error

model1 = LinearRegression()
model2 = KNeighborsRegressor()
model3 = DecisionTreeRegressor()
model4 = RandomForestRegressor()

model_pipeline = [model1, model2, model3, model4]
model_names = ['Linear Regression', 'KNN', 'Decision Tree', 'Random Forest']
scores = {}
scores_rmse={}
scores_mae={}
val_scores= {}
predictions = {}

i=0
for model in model_pipeline:
    mean_score = np.mean(cross_val_score(model, X_train, y_train, cv=5))
    model.fit(X_train, y_train)
    scores[model_names[i]] = mean_score
    prediction=model.predict(X_test)
    predictions[model_names[i]] = prediction
    scores_rmse[model_names[i]] = math.sqrt(mean_squared_error(y_test, prediction))
    scores_mae[model_names[i]] = mean_absolute_error(y_test, prediction)
    val_scores[model_names[i]] = model.score(X_test,y_test)
    
    i = i+1
print('scores: ',scores)
print('cross validation scores: ',val_scores)
print('rmse scores: ',scores_rmse)
print('mae scores: ' ,scores_mae)

#Random Forest seems like the best model...

test_data = pd.DataFrame(y_test)

test_data['RF_pred'] = predictions['Random Forest']
test_data['DT_pred'] = predictions['Decision Tree']
test_data['LR_pred'] = predictions['Linear Regression']
test_data['KNN_pred'] = predictions['KNN']
test_data['models_mean'] = test_data[['RF_pred','DT_pred','KNN_pred']].mean(axis=1)


test_data['winner'] = [y1[i] for i in test_data.index]
test_data['candidate'] = [data['candidate'][i] for i in test_data.index]
test_data['state'] = [data['state'][i] for i in test_data.index]
test_data = test_data[['candidate', 'state', 'percent_popular_vote', 'RF_pred', 'DT_pred', 'KNN_pred','models_mean', 'winner' ]]

test_data.head(60)

#feature importance
importance = model3.feature_importances_
for i,v in enumerate(importance):
    print('Feature: %0d, Score: %.5f' % (i,v))

The most important features are % total tweets (Score: 0.54113), Republican (Score: 0.14864), and % positive tweet Score: (0.03585).

The Linear Regression model was a horrible fit, with negative scores. The Random Forest model 
had the highest score of .92 and lowest RMSE of 6.46. But it only predicted the winner of elections correctly 23 of 36 times.

The Decision Tree model predicted the winner correctly 29 of 36 times.

The KNN model predicted 24 winners correctly.
